CgiSocketクラスの定義を分離
